+++ lib/src/in_app_webview/in_app_webview_settings.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_inappwebview/src/types/user_preferred_content_mode.dart';
 import 'package:flutter_inappwebview_internal_annotations/flutter_inappwebview_internal_annotations.dart';
+import 'dart:typed_data';
 
 import '../android/webview_asset_loader.dart';
 import '../types/action_mode_menu_item.dart';